Add trait UsePusherChannelsNames
